+++ src/components/features/binary/binary.tsx
@@ -1,6 +1,6 @@
 import React, { FunctionComponent } from "react";
 
-import { BinaryFeature, FeatureAccessMode } from "../../../types";
+import { BinaryFeature, Endpoint, FeatureAccessMode } from "../../../types";
 import Toggle from "../../toggle";
 
 import { BaseFeatureProps, BaseViewer, NoAccessError } from "../base";
@@ -11,7 +11,7 @@ const Binary: FunctionComponent<BinaryProps> = (props) => {
   const { feature: { access, endpoint, name, property, value_off: valueOff, value_on: valueOn }, deviceState, onChange } = props;
   if (access & FeatureAccessMode.ACCESS_WRITE) {
     return <Toggle
-      onChange={(value) => onChange(endpoint, { [property]: value })}
+      onChange={(value) => onChange(endpoint as Endpoint, { [property]: value })}
       value={deviceState[property]}
       valueOn={valueOn}
       valueOff={valueOff}