+++ src/components/range-editor/range-editor.tsx
@@ -23,7 +23,7 @@ const RangeEditor: FunctionComponent<RangeProps & Omit<InputHTMLAttributes<HTMLI
 
   const showRange = min !== undefined && max !== undefined;
   return <div className="input-group align-items-center">
-    {steps ? <EnumEditor values={steps} onChange={onChange} /> : null}
+    {steps ? <EnumEditor values={steps} onChange={onChange} value={currentValue} /> : null}
     {showRange ? <input
       id={id}
       min={min}
@@ -42,7 +42,7 @@ const RangeEditor: FunctionComponent<RangeProps & Omit<InputHTMLAttributes<HTMLI
       onChange={e => setCurrentValue(e.target.valueAsNumber)}
       onBlur={() => onChange(currentValue)}
       {...rest}
-      style={showRange ? { 'maxWidth': '100px' } : null}
+      style={showRange ? { 'maxWidth': '100px' } : {}}
     />
     {unit ? <span className="input-group-text" style={{ 'minWidth': '66px' }}>{unit}</span> : null}
   </div>