+++ src/utils.ts
@@ -1,13 +1,15 @@
-import { Device, Endpoint } from './types';
-import { GraphI, LinkI, NodeI } from './components/map/types';
-import { format, TDate } from 'timeago.js';
-import { Group } from './store';
 
-export const genDeviceDetailsLink = (deviceIdentifier: string | number): string => `/device/${deviceIdentifier}`;
+import { Device, Endpoint } from "./types";
+import { GraphI, LinkI, NodeI } from "./components/map/types";
+import { format, TDate } from 'timeago.js';
+import { Group } from "./store";
+
+export const genDeviceDetailsLink = (deviceIdentifier: string | number): string => (`/device/${deviceIdentifier}`);
 
 export const toHex = (input: number, padding = 4): string => {
-    return `0x${`${'0'.repeat(padding)}${input.toString(16)}`.substr(-1 * padding).toUpperCase()}`;
+    return `0x${(`${'0'.repeat(padding)}${input.toString(16)}`).substr(-1 * padding).toUpperCase()}`;
 };
+
 
 export const toHHMMSS = (secs: number): string => {
     if (!secs) {
@@ -18,9 +20,9 @@ export const toHHMMSS = (secs: number): string => {
     const seconds = Math.floor(secs % 60);
 
     return [hours, minutes, seconds]
-        .map((v) => (v < 10 ? `0${v}` : v))
-        .filter((v, i) => v !== '00' || i > 0)
-        .join(':');
+        .map(v => v < 10 ? `0${v}` : v)
+        .filter((v, i) => v !== "00" || i > 0)
+        .join(":");
 };
 
 export type CallbackHandler<T> = (err: boolean, res: T) => void;
@@ -29,59 +31,49 @@ export interface ApiResponse<T> {
     result: T;
 }
 
-export const lastSeen = (lastSeen: string | number, elapsed?: number): string => {
+export const lastSeen = (lastSeen?: string | number, elapsed?: number): string => {
     if (!lastSeen && !elapsed) {
-        return 'N/A';
+        return "N/A";
     }
     let diff: TDate;
     if (elapsed !== undefined) {
         diff = Date.now() - elapsed;
     } else {
-        if (typeof lastSeen === 'string') {
+        if (typeof lastSeen === "string") {
             diff = Date.parse(lastSeen);
         } else {
-            diff = new Date(lastSeen);
+            diff = new Date(lastSeen as number);
         }
     }
     return format(diff);
 };
 
+
 export const sanitizeGraph = (inGraph: GraphI): GraphI => {
     const nodes = {};
     const links = new Map<string, any>();
 
-    inGraph.nodes.forEach((node) => {
+    inGraph.nodes.forEach(node => {
         nodes[node.ieeeAddr] = node;
     });
 
-    inGraph.links
-        .sort((a, b) => a.relationship - b.relationship)
-        .forEach((link) => {
-            const src: NodeI = nodes[link.source.ieeeAddr];
-            const dst: NodeI = nodes[link.target.ieeeAddr];
-            if (src && dst) {
-                const linkId = [link.source.ieeeAddr, link.target.ieeeAddr].sort().join('');
-                const repeatedLink = links.get(linkId);
-                const linkType = [src.type, dst.type].join('2');
-                if (repeatedLink) {
-                    repeatedLink.linkqualities.push(link.linkquality);
-                    repeatedLink.relationships.push(link.relationship);
-                } else {
-                    links.set(linkId, {
-                        ...link,
-                        ...{
-                            source: link.source.ieeeAddr,
-                            linkType,
-                            target: link.target.ieeeAddr,
-                            linkqualities: [link.linkquality],
-                            relationships: [link.relationship],
-                        },
-                    });
-                }
+    inGraph.links.sort((a, b) => a.relationship - b.relationship).forEach(link => {
+        const src: NodeI = nodes[link.source.ieeeAddr];
+        const dst: NodeI = nodes[link.target.ieeeAddr];
+        if (src && dst) {
+            const linkId = [link.source.ieeeAddr, link.target.ieeeAddr].sort().join('');
+            const repeatedLink = links.get(linkId);
+            const linkType = [src.type, dst.type].join('2');
+            if (repeatedLink) {
+                repeatedLink.linkqualities.push(link.linkquality);
+                repeatedLink.relationships.push(link.relationship);
             } else {
-                console.warn('Broken link', link);
+                links.set(linkId, { ...link, ...{ source: link.source.ieeeAddr, linkType, target: link.target.ieeeAddr, linkqualities: [link.linkquality], relationships: [link.relationship]} });
             }
-        });
+        } else {
+            console.warn("Broken link", link);
+        }
+    });
 
     inGraph.links = Array.from(links.values());
     return inGraph;
@@ -95,6 +87,7 @@ export const randomString = (len: number): string => Math.random().toString(36).
 
 export const isSecurePage = (): boolean => location.protocol === 'https:';
 
+
 export const scale = (inputY: number, yRange: Array<number>, xRange: Array<number>): number => {
     const [xMin, xMax] = xRange;
     const [yMin, yMax] = yRange;
@@ -104,6 +97,9 @@ export const scale = (inputY: number, yRange: Array<number>, xRange: Array<numbe
 
     return outputX;
 };
+
+
+
 
 function replacer(key: string, value: object) {
     const originalObject = this[key];
@@ -117,7 +113,7 @@ function replacer(key: string, value: object) {
     }
 }
 
-function reviver(key: string, value: { dataType: string; value: Iterable<readonly [unknown, unknown]> }) {
+function reviver(key: string, value: { dataType: string; value: Iterable<readonly [unknown, unknown]>; }) {
     if (typeof value === 'object' && value !== null) {
         if (value.dataType === 'Map') {
             return new Map(value.value);
@@ -128,11 +124,12 @@ function reviver(key: string, value: { dataType: string; value: Iterable<readonl
 
 export const serialize = (data: object) => {
     return JSON.stringify(data, replacer);
-};
+}
 
 export const deSerialize = (str: string) => {
     return JSON.parse(str, reviver);
-};
+}
+
 
 export const download = (data: object, filename: string): void => {
     const blob = new Blob([serialize(data)], { type: 'octet/stream' });
@@ -147,9 +144,11 @@ export const download = (data: object, filename: string): void => {
         window.URL.revokeObjectURL(url);
         document.body.removeChild(a);
     }, 1);
-};
 
-export const sanitizeZ2MDeviceName = (deviceName: string): string => deviceName?.replace(/:|\s|\//g, '-');
+}
+
+export const sanitizeZ2MDeviceName = (deviceName?: string): string => deviceName ? deviceName.replace(/:|\s|\//g, "-") : "NA";
+
 
 export const getEndpoints = (obj: Device | Group): Endpoint[] => {
     if (!obj) {
@@ -157,7 +156,7 @@ export const getEndpoints = (obj: Device | Group): Endpoint[] => {
     } else if ((obj as Device).endpoints) {
         return Array.from((obj as Device).endpoints.keys());
     } else if ((obj as Group).members) {
-        return (obj as Group).members.map((g) => g.endpoint);
+        return (obj as Group).members.map(g => g.endpoint);
     }
     return [];
-};
+}